% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastis.R
\docType{data}
\name{accipitridaeBasicPastis}
\alias{accipitridaeBasicPastis}
\title{Accipitridae basic data}
\usage{
data(accipitridaeBasicPastis)
}
\description{
Constraint tree and taxon list for Accipitridae
}
\author{
Gavin Thomas, 2013-07-22
}
