% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastis.R
\docType{data}
\name{accipitridaeFullPastis}
\alias{accipitridaeFullPastis}
\title{Accipitridae full data}
\usage{
data(accipitridaeFullPastis)
}
\description{
Constraint tree, taxon list, missing clades, sequence data and template for Accipitridae
}
\author{
Gavin Thomas, 2013-07-22
}
