\name{get_clades}
\alias{get_clades}
\title{Returns the clades for a set of taxa}
\usage{
  get_clades(set, clades)
}
\arguments{
  \item{set}{A vector of taxa}

  \item{clades}{A list of clades, each element contains a
  vector of taxa belonging to that clade}
}
\value{
  A vector of clades represented by the taxa
}
\description{
  This function determines which clades are represented by
  a set of taxa
}

