% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastis-package.R
\docType{package}
\name{pastis-package}
\alias{pastis}
\alias{pastis-package}
\title{pastis: Phylogenetic Assembly with Soft Taxonomic Inferences}
\description{
A pre-processor for mrBayes that assimilates sequences, taxonomic information and tree constraints as per xxx. The main functions of interest for most users will be pastis_simple, pastis_main and conch. The main analysis is conducted with pastis_simple or pastis_main followed by a manual execution of mrBayes (>3.2). The placement of taxa not contained in the tree constraint can be investigated using conch.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://cran.r-project.org/web/packages/pastis/}
}

}
\author{
\strong{Maintainer}: Gavin Thomas \email{gavin.thomas@sheffield.ac.uk}

}
\keyword{internal}
