% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastis.R
\docType{data}
\name{pastis_data_3}
\alias{pastis_data_3}
\title{Example data 3}
\usage{
data(pastis_data_3)
}
\description{
Constraint tree, taxon and missing clade examples
}
\author{
Gavin Thomas, 2013-07-22
}
