% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastis.R
\docType{data}
\name{pastis_data_3_trees}
\alias{pastis_data_3_trees}
\title{Example output}
\usage{
data(pastis_data_3_trees)
}
\description{
[[1]] Constraint tree and [[2]] distributions of trees produced in MrBayes using pastis_data_3
}
\author{
Gavin Thomas, 2013-07-22
}
